requirements.txt
asgiref==3.2.10
Django==3.0.5
django-appconf==1.0.4
django-imagekit==4.0.2
django-mathfilters==1.0.0
djangorestframework==3.11.1
djongo==1.3.3
pilkit==2.0
Pillow==7.2.0
pymongo==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
